# packages/saas-platform-spec/admin-api.openapi.yaml
openapi: 3.1.0
...
info:
    title: SaaS-Plattform SuperAdmin-API
    version: 0.1.0-draft
    description: |
        Read- und Write-Operationen für die Plattform-Verwaltung:
        Tenants, Users, Subscriptions, Promo-Codes, Audit-Log,
        Manifest-Discovery, plus Plan-Catalog-Lookup.
        Authentifizierung via JWT mit Rolle SUPER_ADMIN; alle schreibenden
        Operationen verlangen zusätzlich einen TOTP-Code im Header
        X-Mfa-Code, sofern der Admin MFA aktiviert hat (sonst 401
        MFA_REQUIRED). Manifest ist nie Security-Quelle (siehe SPEC §10.4).
servers:
    - url: https://{host}/api/v1/admin
      variables:
          host:
              default: localhost:3000
              description: Konsumenten-spezifischer Host (Domain der jeweiligen App)
...
security:
    - bearerAuth: []
paths:
...
    /tenants/{slug}:
        parameters:
            - $ref: '#/components/parameters/TenantSlug'
        get:
            tags: [tenants]
            summary: Detail eines Tenants inkl. Subscription, Add-ons, Quota-Verbrauch
            responses:
                '200':
                    description: Detail
                    content:
                        application/json:
                            schema: { $ref: '#/components/schemas/AdminTenantDetail' }
        delete:
            tags: [tenants]
            summary: Tenant soft-deleten (deletedAt setzen)
            x-mfa-required: true
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            required: [reason, confirmSlug]
                            properties:
                                reason: { type: string, minLength: 10 }
                                confirmSlug:
                                    {
                                        type: string,
                                        description: 'Muss exakt dem Tenant-Slug entsprechen',
                                    }
            responses:
                '200': { description: Gelöscht }
...
    /tenants/{slug}/reactivate:
        parameters:
            - $ref: '#/components/parameters/TenantSlug'
        post:
            tags: [tenants]
            summary: Suspendierten Tenant wieder aktivieren
            x-mfa-required: true
            responses:
                '200': { description: Reaktiviert }
...
    /catalog/bundles/{id}:
        parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
        get:
            tags: [catalog]
            summary: Bundle-Stamm Detail
            responses: { '200': { description: Detail } }
        patch:
            tags: [catalog]
            summary: Bundle-Stamm aktualisieren (Label/Beschreibung/Icon/sortOrder/i18n)
            x-mfa-required: true
            requestBody:
                required: true
                content: { application/json: { schema: { $ref: '#/components/schemas/CatalogBundleUpdate' } } }
            responses: { '200': { description: Aktualisiert } }
        delete:
            tags: [catalog]
            summary: Bundle-Stamm soft-deleten
            x-mfa-required: true
            responses: { '200': { description: Gelöscht } }
    /catalog/bundles/{id}/versions:
        parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
        get:
            tags: [catalog]
            summary: BundleVersionen auflisten
            responses: { '200': { description: Liste } }
        post:
            tags: [catalog]
            summary: BundleVersion-Entwurf anlegen (unpublished)
            x-mfa-required: true
            requestBody:
                required: true
                content: { application/json: { schema: { $ref: '#/components/schemas/CatalogBundleVersionDraft' } } }
            responses: { '201': { description: Entwurf angelegt } }
...
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
    parameters:
...
        TenantSlug:
            name: slug
            in: path
            required: true
            schema: { type: string, pattern: '^[a-z0-9][a-z0-9-]{1,62}[a-z0-9]$' }
    schemas:
...
        CatalogBundleVersionDraft:
            type: object
            required: [features]
            properties:
                features:
                    type: array
                    items: { type: string, pattern: '^[A-Z][A-Z0-9_]*$' }
                quotas:
                    type: object
                    additionalProperties: { type: number }
                compatibility:
                    type: object
                    description: 'beide leer/weggelassen = frei buchbar überall (E5)'
                    properties:
                        businessTypeKeys: { type: array, items: { type: string } }
                        planIds: { type: array, items: { type: string } }
                monthlyNet:
                    { type: string, pattern: '^\d+(\.\d{1,2})?$', nullable: true, description: 'EUR-String; null = Override-Resolution' }
                yearlyNet: { type: string, pattern: '^\d+(\.\d{1,2})?$', nullable: true }
                marketed: { type: boolean }
                changeNote: { type: string, maxLength: 2000 }
                baseVersionId: { type: string, nullable: true }
                validFrom: { type: string, nullable: true }
                validUntil: { type: string, nullable: true }
...
        CatalogBundleUpdate:
            type: object
            properties:
                label: { type: string, minLength: 1, maxLength: 120 }
                description: { type: string, maxLength: 2000, nullable: true }
                icon: { type: string, maxLength: 64, nullable: true }
                sortOrder: { type: integer, minimum: 0, maximum: 10000 }
                i18n: { type: object, additionalProperties: { type: object } }
...
        AdminTenantRow:
            type: object
            properties:
                id: { type: string }
                slug: { type: string }
                name: { type: string }
                plan: { type: string }
                status: { type: string }
                createdAt: { type: string, format: date-time }
                usage:
                    type: object
                    additionalProperties: { type: integer }
                    description: 'Aktueller Verbrauch je quotaKey'
        AdminTenantDetail:
            allOf:
                - $ref: '#/components/schemas/AdminTenantRow'
                - type: object
                  properties:
                      email: { type: string, format: email }
                      taxId: { type: string }
                      vatId: { type: string }
                      address: { type: object }
                      settings: { type: object }
                      subscription: { $ref: '#/components/schemas/SubscriptionDetail' }
                      recentAudit:
                          type: array
                          items: { $ref: '#/components/schemas/AuditEntryDto' }
...
        SubscriptionDetail:
            type: object
            properties:
                id: { type: string }
                planId: { type: string }
                billingCycle: { type: string }
                status: { type: string }
                trialEndsAt: { type: string, format: date-time }
                startedAt: { type: string, format: date-time }
                canceledAt: { type: string, format: date-time }
                isPilot: { type: boolean }
                pilotEndsAt: { type: string, format: date-time }
                customLimits:
                    type: object
                    additionalProperties: { type: integer }
                customMonthlyNet: { type: number, format: decimal, nullable: true }
                effectiveLimits:
                    type: object
                    additionalProperties: { type: integer }
                    description: 'Plan-Quotas + Bundle-Beiträge + customLimits-Overrides'
                effectiveFeatures:
                    type: array
                    items: { type: string }
                promoRedemption:
                    $ref: '#/components/schemas/PromoRedemptionDto'

        # ────────── User ──────────
...
        PromoRedemptionDto:
            type: object
            properties:
                id: { type: string }
                promoCodeId: { type: string }
                promoCode: { type: string, description: 'Convenience: der Code-String' }
                subscriptionId: { type: string }
                tenantSlug: { type: string }
                appliedValueType: { type: string }
                appliedValue: { type: number, format: decimal }
                startsAt: { type: string, format: date-time }
                endsAt: { type: string, format: date-time, nullable: true }
                status: { type: string }
                redeemedAt: { type: string, format: date-time }
                reversedAt: { type: string, format: date-time, nullable: true }

        # ────────── Audit ──────────
        AuditEntryDto:
            type: object
            properties:
                id: { type: string }
                tenantId: { type: string, nullable: true }
                userId: { type: string, nullable: true }
                userEmail: { type: string, nullable: true, description: 'Convenience-Lookup' }
                entity: { type: string }
                entityId: { type: string }
                action: { type: string }
                changes: { type: object, additionalProperties: true }
                actorTag: { type: string, nullable: true }
                ipAddress: { type: string, nullable: true }
                createdAt: { type: string, format: date-time }

        # ────────── Dashboard ──────────
...